+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/menus/options/events/NewGameMenuOptionEvent.java
@@ -12,7 +12,7 @@ public class NewGameMenuOptionEvent extends EnterStateMenuOptionEvent {
     private StringMenuOptionValue playerName;
 
     public NewGameMenuOptionEvent(StateBasedGame game, int stateId, MenuState menu, StringMenuOptionValue playerName) {
-        super(game, stateId);
+        super(game, menu, stateId);
 
         this.menu = menu;
         this.playerName = playerName;