+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [@mjkuranda](https://github.com/mjkuranda): HighScoreHandler to handle high score.
 - [@mjkuranda](https://github.com/mjkuranda): Saving and writing to the file of high scores.
 
+### Fixed
+- [@mjkuranda](https://github.com/mjkuranda): Resetting the other menu states after entering to another one.
+
 ## [0.17.0] - 2023-08-03
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Loading state.